requirements.txt
asgiref==3.6.0
attrs==20.3.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
check50==3.3.7
chromedriver-py==115.0.5790.170
cryptography==39.0.2
distro==1.8.0
Django==4.2.1
docker==6.1.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
jellyfish==0.9.0
Jinja2==3.1.2
jsonschema==3.2.0
lib50==3.0.5
markdown2==2.4.8
MarkupSafe==2.1.2
numpy==1.25.2
outcome==1.2.0
packaging==23.1
paramiko==3.2.0
pexpect==4.8.0
pipreqs==0.4.13
ptyprocess==0.7.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==5.4.1
requests==2.28.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
submit50==3.1.1
termcolor==1.1.0
texttable==1.6.7
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.15
webdriver-manager==4.0.0
websocket-client==0.59.0
wsproto==1.2.0
yarg==0.1.9
